requirements.txt
pandas==2.0.3
numpy==1.24.3
sqlalchemy==2.0.19
psycopg2-binary==2.9.6
python-dotenv==1.0.0
lxml==4.9.3
streamlit==1.31.0
plotly==5.18.0
pytest==8.4.2